MutableTree: Refactor support for versions
LoadVersionFast is changed so that passing a zero version works
correctly by obtaining the latest version via the nodedb.

Load is changed so that it uses LoadVersionFast to avoid traversing all
the versions in the database. To avoid issues with tree.versions, this
internal map is removed and replaced with a hasRoot query against the
nodedb.